Fix call to extractAll function
This patch is partially necessary to correct the `--extract-all` flag.
With this correction, `--extract-all` will work if there were no
directories ingested from an input zip.

Issue 15 reports on the problem with directories from an ingested zip.
https://github.com/aff4/pyaff4/issues/15

This patch builds on Pull Request 13, as I'd also found the binary-
output mode was necessary, though in a different spot.
https://github.com/aff4/pyaff4/pull/13

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>