Fix extraction of Windows absolute paths (#25)
* Fix extraction of Windows absolute paths

os.path.join() discards earlier arguments if a part contains an absolute path, i.e. starts with slash(es) or a drive letter. As a result, in extractFromVolume and extractAllFromVolume, destFolder is ignored if pathName is a Windows absolute path. This commit strips leading slashes and removes the colon if a drive letter is present in pathName.

* Update aff4.py to print skipped paths

Fixed print statement in addPathNamesToVolume to include the paths of files being skipped (instead of "%s").

* Fix comparison in data_store.py